mask_rcnn_onnx/src/requirements.txt
pillow
numpy
onnx
onnxruntime
Cython
pycocotools
opencv-python